functions/ingest_csv/requirements.txt
google-cloud-bigquery>=3.23.0
google-cloud-storage>=2.16.0
